Extract ad message label into its own component
